move: test class 패키지 이동
