+++ src/test/java/blackjack/domain/game/BlackjackHandTest.java
@@ -1,6 +1,8 @@
-package blackjack.domain.card;
+package blackjack.domain.game;
 
-import blackjack.domain.game.BlackjackHand;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Symbol;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
