+++ src/test/java/blackjack/domain/participant/NameTest.java
@@ -1,6 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
-import blackjack.domain.participant.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
